tools/requirements.txt
langchain
langchain-community
langchain-openai
chromadb
openai
python-dotenv
pypdf
tiktoken
